+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/method_channel_document_change.dart
@@ -3,26 +3,23 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
 
-import 'utils/maps.dart';
-
-/// A DocumentChange represents a change to the documents matching a query.
-///
-/// It contains the document affected and the type of change that occurred
-/// (added, modified, or removed).
+/// An implementation of [DocumentChangePlatform] that uses [MethodChannel] to
+/// communicate with Firebase plugins.
 class MethodChannelDocumentChange extends DocumentChangePlatform {
-  /// Create instance of [MethodChannelDocumentChange] using [data]
+  /// Creates a [MethodChannelDocumentChange] from the given [data]
   MethodChannelDocumentChange(
-      Map<dynamic, dynamic> data, FirestorePlatform firestore)
+      FirebaseFirestorePlatform firestore, Map<String, dynamic> data)
       : super(DocumentChangeType.values.firstWhere((DocumentChangeType type) {
           return type.toString() == data['type'];
         }),
             data['oldIndex'],
             data['newIndex'],
             DocumentSnapshotPlatform(
-              data['path'],
-              asStringKeyedMap(data['document']),
-              SnapshotMetadataPlatform(data['metadata']['hasPendingWrites'],
-                  data['metadata']['isFromCache']),
               firestore,
+              data['path'],
+              <String, dynamic>{
+                'data': Map<String, dynamic>.from(data['data']),
+                'metadata': Map<String, dynamic>.from(data['metadata']),
+              },
             ));
 }