+++ packages/firebase_ml_vision/CHANGELOG.md
@@ -158,7 +158,7 @@ You should now call `detector.close()` when a detector will no longer be used.
 * **Breaking Change** Removed on-device model dependencies from plugin.
   `Android` now requires adding the on-device label detector dependency manually.
   `iOS` now requires adding the on-device barcode/face/label/text detector dependencies manually.
-  See the `README.md` for more details. https://pub.dartlang.org/packages/firebase_ml_vision#-readme-tab-
+  See the `README.md` for more details. https://pub.dev/packages/firebase_ml_vision#-readme-tab-
 
 ## 0.5.1+2